requirements.txt
Flask
Pillow
torch
torchvision
transformers